--- src/Components/AdSubmitForm/Forms/Bachelor.js
@@ -1,13 +1,0 @@
-import React from 'react';
-
-const Bachelor = () => {
-
-
-    return(
-        <div>
-            abc
-        </div>
-    )
-}
-
-export default Bachelor
