+++ src/Components/configs.js
@@ -1,0 +1,1 @@
+export const BASE_API_URL = 'https://to-let-back.herokuapp.com/api'
