+++ src/redux/Constants.js
@@ -25,6 +25,8 @@ export const CHANGE_IMAGE_4 = 'CHANGE_IMAGE_4'
 
 //adsubmit
 
-export const SUBMIT_NEW_ADD = 'SUBMIT_NEW_ADD'
+export const SUBMIT_NEW_ADD_REQUEST = 'SUBMIT_NEW_ADD_REQUEST'
+export const SUBMIT_NEW_ADD_SUCCESS = 'SUBMIT_NEW_ADD_SUCCESS'
+export const SUBMIT_NEW_ADD_FAILED = 'SUBMIT_NEW_ADD_FAILED'
